make generic edit screen
